Added pipeline helper fumctions
