+++ basic_pipelines/detection.py
@@ -2,26 +2,15 @@ import gi
 gi.require_version('Gst', '1.0')
 from gi.repository import Gst, GLib
 import os
-import argparse
-import multiprocessing
 import numpy as np
-import setproctitle
 import cv2
-import time
 import hailo
 from hailo_rpi_common import (
-    get_default_parser,
-    QUEUE,
-    SOURCE_PIPELINE,
-    DETECTION_PIPELINE,
-    INFERENCE_PIPELINE_WRAPPER,
-    USER_CALLBACK_PIPELINE,
-    DISPLAY_PIPELINE,
     get_caps_from_pad,
     get_numpy_from_buffer,
-    GStreamerApp,
     app_callback_class,
 )
+from detection_pipeline import GStreamerDetectionApp
 
 # -----------------------------------------------------------------------------------------------
 # User-defined class to be used in the callback function
@@ -87,87 +76,8 @@ def app_callback(pad, info, user_data):
     print(string_to_print)
     return Gst.PadProbeReturn.OK
 
-
-# -----------------------------------------------------------------------------------------------
-# User Gstreamer Application
-# -----------------------------------------------------------------------------------------------
-
-# This class inherits from the hailo_rpi_common.GStreamerApp class
-class GStreamerDetectionApp(GStreamerApp):
-    def __init__(self, args, user_data):
-        # Call the parent class constructor
-        super().__init__(args, user_data)
-        # Additional initialization code can be added here
-        # Set Hailo parameters these parameters should be set based on the model used
-        self.batch_size = 2
-        self.network_width = 640
-        self.network_height = 640
-        self.network_format = "RGB"
-        nms_score_threshold = 0.3
-        nms_iou_threshold = 0.45
-
-        if args.hef_path is not None:
-            self.hef_path = args.hef_path
-        # Set the HEF file path based on the network
-        elif args.network == "yolov6n":
-            self.hef_path = os.path.join(self.current_path, '../resources/yolov6n.hef')
-        elif args.network == "yolov8s":
-            self.hef_path = os.path.join(self.current_path, '../resources/yolov8s_h8l.hef')
-        else:
-            assert False, "Invalid network type"
-
-        # User-defined label JSON file
-        self.labels_json = args.labels_json
-
-        self.app_callback = app_callback
-
-        self.thresholds_str = (
-            f"nms-score-threshold={nms_score_threshold} "
-            f"nms-iou-threshold={nms_iou_threshold} "
-            f"output-format-type=HAILO_FORMAT_TYPE_FLOAT32"
-        )
-
-        # Set the process title
-        setproctitle.setproctitle("Hailo Detection App")
-
-        self.create_pipeline()
-
-    def get_pipeline_string(self):
-        source_pipeline = SOURCE_PIPELINE(self.video_source)
-        detection_pipeline = DETECTION_PIPELINE(hef_path=self.hef_path, batch_size=self.batch_size, labels_json=self.labels_json, additional_params=self.thresholds_str)
-        detection_pipeline_wrapper = INFERENCE_PIPELINE_WRAPPER(inner_pipeline=detection_pipeline)
-        user_callback_pipeline = USER_CALLBACK_PIPELINE()
-        display_pipeline = DISPLAY_PIPELINE(video_sink=self.video_sink, sync=self.sync, show_fps=self.show_fps)
-        pipeline_string = (
-            f'{source_pipeline} '
-            f'{detection_pipeline_wrapper} ! '
-            f'{user_callback_pipeline} ! '
-            f'{display_pipeline}'
-        )
-        print(pipeline_string)
-        return pipeline_string
-
 if __name__ == "__main__":
     # Create an instance of the user app callback class
     user_data = user_app_callback_class()
-    parser = get_default_parser()
-    # Add additional arguments here
-    parser.add_argument(
-        "--network",
-        default="yolov6n",
-        choices=['yolov6n', 'yolov8s'],
-        help="Which Network to use, default is yolov6n",
-    )
-    parser.add_argument(
-        "--hef-path",
-        default=None,
-        help="Path to HEF file",
-    )
-    parser.add_argument(
-        "--labels-json",
-        default=None,
-        help="Path to costume labels JSON file",
-    )
-    args = parser.parse_args()
-    app = GStreamerDetectionApp(args, user_data)
+    app = GStreamerDetectionApp(app_callback, user_data)
     app.run()
