+++ basic_pipelines/instance_segmentation_pipeline.py
@@ -1,0 +1,76 @@
+import gi
+gi.require_version('Gst', '1.0')
+from gi.repository import Gst, GLib
+import os
+import argparse
+import multiprocessing
+import numpy as np
+import setproctitle
+import cv2
+import time
+import hailo
+from hailo_rpi_common import (
+    get_default_parser,
+    QUEUE,
+    SOURCE_PIPELINE,
+    INFERENCE_PIPELINE,
+    USER_CALLBACK_PIPELINE,
+    DISPLAY_PIPELINE,
+    get_caps_from_pad,
+    get_numpy_from_buffer,
+    GStreamerApp,
+    app_callback_class,
+    dummy_callback,
+)
+
+#-----------------------------------------------------------------------------------------------
+# User Gstreamer Application
+# -----------------------------------------------------------------------------------------------
+
+# This class inherits from the hailo_rpi_common.GStreamerApp class
+
+class GStreamerInstanceSegmentationApp(GStreamerApp):
+    def __init__(self, app_callback, user_data):
+        parser = get_default_parser()
+        args = parser.parse_args()
+        # Call the parent class constructor
+        super().__init__(args, user_data)
+        # Additional initialization code can be added here
+        # Set Hailo parameters these parameters should be set based on the model used
+        self.batch_size = 2
+        self.network_width = 640
+        self.network_height = 640
+        self.network_format = "RGB"
+        self.default_post_process_so = os.path.join(self.postprocess_dir, 'libyolov5seg_post.so')
+        self.post_function_name = "yolov5seg"
+        self.hef_path = os.path.join(self.current_path, '../resources/yolov5n_seg_h8l_mz.hef')
+        self.app_callback = app_callback
+
+        # Set the process title
+        setproctitle.setproctitle("Hailo Instance Segmentation App")
+
+        self.create_pipeline()
+
+    def get_pipeline_string(self):
+        source_pipeline = SOURCE_PIPELINE(video_source=self.video_source)
+        infer_pipeline = INFERENCE_PIPELINE(
+            hef_path=self.hef_path,
+            post_process_so=self.default_post_process_so,
+            post_function_name=self.post_function_name
+        )
+        user_callback_pipeline = USER_CALLBACK_PIPELINE()
+        display_pipeline = DISPLAY_PIPELINE(video_sink=self.video_sink, sync=self.sync, show_fps=self.show_fps)
+        pipeline_string = (
+            f'{source_pipeline} '
+            f'{infer_pipeline} ! '
+            f'{user_callback_pipeline} ! '
+            f'{display_pipeline}'
+        )
+        print(pipeline_string)
+        return pipeline_string
+
+if __name__ == "__main__":
+    # Create an instance of the user app callback class
+    user_data = app_callback_class()
+    app = GStreamerInstanceSegmentationApp(dummy_callback, user_data)
+    app.run()
