+++ basic_pipelines/pose_estimation.py
@@ -2,25 +2,15 @@ import gi
 gi.require_version('Gst', '1.0')
 from gi.repository import Gst, GLib
 import os
-import argparse
-import multiprocessing
 import numpy as np
-import setproctitle
 import cv2
-import time
 import hailo
 from hailo_rpi_common import (
-    get_default_parser,
-    QUEUE,
-    SOURCE_PIPELINE,
-    INFERENCE_PIPELINE,
-    USER_CALLBACK_PIPELINE,
-    DISPLAY_PIPELINE,
     get_caps_from_pad,
     get_numpy_from_buffer,
-    GStreamerApp,
     app_callback_class,
 )
+from pose_estimation_pipeline import GStreamerPoseEstimationApp
 
 # -----------------------------------------------------------------------------------------------
 # User-defined class to be used in the callback function
@@ -59,6 +49,9 @@ def app_callback(pad, info, user_data):
     roi = hailo.get_roi_from_buffer(buffer)
     detections = roi.get_objects_typed(hailo.HAILO_DETECTION)
 
+    # Get the keypoints
+    keypoints = get_keypoints()
+
     # Parse the detections
     for detection in detections:
         label = detection.get_label()
@@ -70,19 +63,14 @@ def app_callback(pad, info, user_data):
             landmarks = detection.get_objects_typed(hailo.HAILO_LANDMARKS)
             if len(landmarks) != 0:
                 points = landmarks[0].get_points()
-                left_eye = points[1]  # assuming 1 is the index for the left eye
-                right_eye = points[2]  # assuming 2 is the index for the right eye
-                # The landmarks are normalized to the bounding box, we also need to convert them to the frame size
-                left_eye_x = int((left_eye.x() * bbox.width() + bbox.xmin()) * width)
-                left_eye_y = int((left_eye.y() * bbox.height() + bbox.ymin()) * height)
-                right_eye_x = int((right_eye.x() * bbox.width() + bbox.xmin()) * width)
-                right_eye_y = int((right_eye.y() * bbox.height() + bbox.ymin()) * height)
-                string_to_print += (f" Left eye: x: {left_eye_x:.2f} y: {left_eye_y:.2f} Right eye: x: {right_eye_x:.2f} y: {right_eye_y:.2f}\n")
-                if user_data.use_frame:
-                    # Add markers to the frame to show eye landmarks
-                    cv2.circle(frame, (left_eye_x, left_eye_y), 5, (0, 255, 0), -1)
-                    cv2.circle(frame, (right_eye_x, right_eye_y), 5, (0, 255, 0), -1)
-                    # Note: using imshow will not work here, as the callback function is not running in the main thread
+                for eye in ['left_eye', 'right_eye']:
+                    keypoint_index = keypoints[eye]
+                    point = points[keypoint_index]
+                    x = int((point.x() * bbox.width() + bbox.xmin()) * width)
+                    y = int((point.y() * bbox.height() + bbox.ymin()) * height)
+                    string_to_print += f"{eye}: x: {x:.2f} y: {y:.2f}\n"
+                    if user_data.use_frame:
+                        cv2.circle(frame, (x, y), 5, (0, 255, 0), -1)
 
     if user_data.use_frame:
         # Convert the frame to BGR
@@ -92,80 +80,33 @@ def app_callback(pad, info, user_data):
     print(string_to_print)
     return Gst.PadProbeReturn.OK
 
-
-
 # This function can be used to get the COCO keypoints coorespondence map
 def get_keypoints():
     """Get the COCO keypoints and their left/right flip coorespondence map."""
     keypoints = {
-        'nose': 1,
-        'left_eye': 2,
-        'right_eye': 3,
-        'left_ear': 4,
-        'right_ear': 5,
-        'left_shoulder': 6,
-        'right_shoulder': 7,
-        'left_elbow': 8,
-        'right_elbow': 9,
-        'left_wrist': 10,
-        'right_wrist': 11,
-        'left_hip': 12,
-        'right_hip': 13,
-        'left_knee': 14,
-        'right_knee': 15,
-        'left_ankle': 16,
-        'right_ankle': 17,
+        'nose': 0,
+        'left_eye': 1,
+        'right_eye': 2,
+        'left_ear': 3,
+        'right_ear': 4,
+        'left_shoulder': 5,
+        'right_shoulder': 6,
+        'left_elbow': 7,
+        'right_elbow': 8,
+        'left_wrist': 9,
+        'right_wrist': 10,
+        'left_hip': 11,
+        'right_hip': 12,
+        'left_knee': 13,
+        'right_knee': 14,
+        'left_ankle': 15,
+        'right_ankle': 16,
     }
 
     return keypoints
-#-----------------------------------------------------------------------------------------------
-# User Gstreamer Application
-# -----------------------------------------------------------------------------------------------
-
-# This class inherits from the hailo_rpi_common.GStreamerApp class
-
-class GStreamerPoseEstimationApp(GStreamerApp):
-    def __init__(self, args, user_data):
-        # Call the parent class constructor
-        super().__init__(args, user_data)
-        # Additional initialization code can be added here
-        # Set Hailo parameters these parameters should be set based on the model used
-        self.batch_size = 2
-        self.network_width = 640
-        self.network_height = 640
-        self.network_format = "RGB"
-        self.default_post_process_so = os.path.join(self.postprocess_dir, 'libyolov8pose_post.so')
-        self.post_function_name = "filter"
-        self.hef_path = os.path.join(self.current_path, '../resources/yolov8s_pose_h8l_pi.hef')
-        self.app_callback = app_callback
-
-        # Set the process title
-        setproctitle.setproctitle("Hailo Pose Estimation App")
-
-        self.create_pipeline()
-
-    def get_pipeline_string(self):
-        source_pipeline = SOURCE_PIPELINE(video_source=self.video_source)
-        infer_pipeline = INFERENCE_PIPELINE(
-            hef_path=self.hef_path,
-            post_process_so=self.default_post_process_so,
-            post_function_name=self.post_function_name
-        )
-        user_callback_pipeline = USER_CALLBACK_PIPELINE()
-        display_pipeline = DISPLAY_PIPELINE(video_sink=self.video_sink, sync=self.sync, show_fps=self.show_fps)
-        pipeline_string = (
-            f'{source_pipeline} '
-            f'{infer_pipeline} ! '
-            f'{user_callback_pipeline} ! '
-            f'{display_pipeline}'
-        )
-        print(pipeline_string)
-        return pipeline_string
 
 if __name__ == "__main__":
     # Create an instance of the user app callback class
     user_data = user_app_callback_class()
-    parser = get_default_parser()
-    args = parser.parse_args()
-    app = GStreamerPoseEstimationApp(args, user_data)
+    app = GStreamerPoseEstimationApp(app_callback, user_data)
     app.run()
