+++ cpp/meson.build
@@ -8,7 +8,21 @@ yolo_hailortpp_sources = [
 ]
 shared_library('yolo_hailortpp_post',
     yolo_hailortpp_sources,
-    include_directories : rapidjson_inc,  
+    include_directories : rapidjson_inc,
+    dependencies : postprocess_dep,
+    gnu_symbol_visibility : 'default',
+    install: true,
+    install_dir: join_paths(meson.project_source_root(), 'resources'),
+)
+################################################
+# REMOVE LABELS POST SOURCES
+################################################
+
+remove_labels_sources = [
+    'remove_labels.cpp',
+]
+shared_library('remove_labels',
+    remove_labels_sources,
     dependencies : postprocess_dep,
     gnu_symbol_visibility : 'default',
     install: true,
